Moduleによるアイテム絞りこみ部分の切り出し
